fix: implementing adjustments after migration to dotykacka v2 api

+++ src/main/java/pl/grizzlysoftware/chlorek/core/model/InvoiceItem.java
@@ -17,5 +17,5 @@ public class InvoiceItem {
     public BigDecimal grossSellPrice;
     public BigDecimal netPurchaseTotalPrice;
     public BigDecimal currentGrossSellPrice;
-    public Double quantity;
+    public BigDecimal quantity;
 }
